[AIRFLOW-4843] Squash me: Move operator to contrib
