+++ airflow/contrib/operators/swarm_operator.py
